メタデータの解決と保存の処理を統一
